2o5t: Split catalog wrapper into per-service modules for the file-size guard
